+++ lib/screens/addTransaction/selectValue.dart
@@ -1,5 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:your_budget/appState.dart';
+import 'package:your_budget/components/addDialog.dart';
+import 'package:your_budget/models/categories.dart';
 import 'package:your_budget/models/entries.dart';
 import 'package:provider/provider.dart';
 
@@ -16,6 +18,7 @@ class SelectValuePageState extends State<SelectValuePage> {
   TextEditingController searchController = new TextEditingController();
   String filter;
   bool isPayee;
+  bool isSubcategories;
 
   AppState appState;
   @override
@@ -23,6 +26,7 @@ class SelectValuePageState extends State<SelectValuePage> {
     super.initState();
     appState = Provider.of(context, listen: false);
     isPayee = widget.title == "Payees";
+    isSubcategories = widget.title == "Subcategories";
     searchController.addListener(() {
       setState(() {
         filter = searchController.text;
@@ -40,19 +44,94 @@ class SelectValuePageState extends State<SelectValuePage> {
     Navigator.pop(context, item);
   }
 
-  handleCreateNewPayee(String payeeName) async {
+  String newPayeeValidator(String payeeName) {
     if (payeeName == null || payeeName.trim() == "") {
-      print("Name must be valid");
-      //TODO: Add error message
-      return;
+      return "Name must be valid";
     }
-    print("Created payee $payeeName");
-    Payee payee = await appState.addPayee(payeeName: payeeName);
-    handlePopContext(payee);
+    return null;
+  }
+
+  void createNewPayee(
+      {@required BuildContext context, String defaultName}) async {
+    String hint = "Add new payee";
+    String payeeName = await addDialog(
+        context: context,
+        title: hint,
+        hintText: hint,
+        successButtonName: "Create and select",
+        defaultValue: defaultName,
+        nameValidator: newPayeeValidator);
+
+    if (payeeName != null) {
+      print("Created payee $payeeName");
+      Payee payee = await appState.addPayee(payeeName: payeeName);
+      handlePopContext(payee);
+    }
+  }
+
+  List<dynamic> _getDuplicateEntries() {
+    var map = Map<String, int>(); // Mapping between name and id
+
+    Set<int> duplicateIDs = Set<int>();
+    widget.listEntries.forEach((element) {
+      if (element is SubCategory) {
+        if (!map.containsKey(element.name)) {
+          map[element.name] = element.id;
+        } else {
+          duplicateIDs.add(element.id);
+          duplicateIDs.add(map[element.name]); //Other duplicate that is in map
+        }
+      }
+    });
+    return appState.subcategories
+        .where((element) => duplicateIDs.contains(element.id))
+        .toList();
+  }
+
+  List _addLabelForDuplicateEntries(List<dynamic> duplicateEntries) {
+    List modifiedListEntries = [];
+    List maincategories =
+        appState.allCategories.whereType<MainCategory>().toList();
+
+    widget.listEntries.forEach((entry) {
+      if (entry is SubCategory) {
+        bool isDuplicate = duplicateEntries.singleWhere(
+                (duplicate) => duplicate.id == entry.id,
+                orElse: () => null) !=
+            null;
+
+        if (isDuplicate) {
+          var modifiedEntry = _addCategoryName(entry, maincategories);
+          modifiedListEntries.add(modifiedEntry);
+        } else {
+          modifiedListEntries.add(entry);
+        }
+      }
+    });
+    return modifiedListEntries;
+  }
+
+  dynamic _addCategoryName(var entry, List maincategories) {
+    var modifiedEntry = entry.copy();
+    MainCategory category = maincategories
+        .singleWhere((maincat) => maincat.id == entry.parentId, orElse: null);
+    modifiedEntry.name = modifiedEntry.name + ' (' + category.name + ')' ?? "";
+    return modifiedEntry;
   }
 
   @override
   Widget build(BuildContext context) {
+    List listEntries = [];
+
+    if (isSubcategories) {
+      List<dynamic> duplicates = _getDuplicateEntries();
+      if (duplicates.isNotEmpty)
+        listEntries = _addLabelForDuplicateEntries(duplicates);
+      else
+        listEntries = widget.listEntries;
+    } else
+      listEntries = widget.listEntries;
+
     return Scaffold(
       resizeToAvoidBottomInset: false,
       appBar: AppBar(
@@ -60,7 +139,7 @@ class SelectValuePageState extends State<SelectValuePage> {
       ),
       body: Column(
         children: [
-         Padding(
+          Padding(
             padding: EdgeInsets.only(top: 20.0),
           ),
           TextField(
@@ -75,31 +154,35 @@ class SelectValuePageState extends State<SelectValuePage> {
           if (isPayee)
             ListTile(
               title: Text(
-                "Create new payee \"${(filter == null) ? "" : filter}\"",
-                style: TextStyle(fontStyle: FontStyle.italic),
+                "Create new payee",
+                style: TextStyle(
+                    fontStyle: FontStyle.italic, fontWeight: FontWeight.bold),
               ),
-              onTap: () => handleCreateNewPayee(filter),
+              onTap: () => createNewPayee(
+                  context: context, defaultName: searchController.text),
             ),
 
           Expanded(
               child: ListView.separated(
             shrinkWrap: true,
-            itemCount: widget.listEntries.length,
+            itemCount: listEntries.length,
             separatorBuilder: (BuildContext context, int index) =>
                 Divider(height: 1, color: Colors.black12),
             itemBuilder: (BuildContext context, int index) {
-              var item = widget.listEntries[index];
-              print(item);
-              bool noFilter = filter == null || filter == "";
+              var item = listEntries[index];
               var itemToShow = item is Text ? item : Text(item.name);
               var itemToFilter = item is Text ? item.data : item.name;
+              bool noFilter = filter == null || filter == "";
+
               if (noFilter == true)
                 // The filter is empty, show everything
-                return ListTile(title: itemToShow, onTap: () => handlePopContext(item));
-              else if (noFilter = false) {
+                return ListTile(
+                    title: itemToShow, onTap: () => handlePopContext(item));
+              else if (noFilter == false) {
                 // The filter is not empty, we filter by name
                 if (itemToFilter.toLowerCase().contains(filter.toLowerCase()))
-                  return ListTile(title: item, onTap: () => handlePopContext(item));
+                  return ListTile(
+                      title: item, onTap: () => handlePopContext(item));
               }
               // There is an error
               return Container();