+++ lib/screens/showTransactions/components/transactionList.dart
@@ -1,5 +1,8 @@
+import 'dart:collection';
+
 import 'package:flutter/material.dart';
 import 'package:your_budget/appState.dart';
+import 'package:your_budget/models/constants.dart';
 import 'package:your_budget/models/entries.dart';
 import 'package:your_budget/screens/showTransactions/components/transactionRow.dart';
 
@@ -25,15 +28,15 @@ class _TransactionListState extends State<TransactionList> {
 
   @override
   Widget build(BuildContext context) {
-    List<MoneyTransaction> transactionsOfAccount = widget.appState.transactions
-        .where((transaction) => transaction.accountID == this.widget.account.id)
-        .toList();
+    List<MoneyTransaction> transactionsOfAccount =
+        _getMoneyTransactions(
+            widget.appState.transactions, this.widget.account.id);
 
-    return new Container(
+    return Container(
         child: Scrollbar(
       isAlwaysShown: true,
       controller: _scrollController,
-      child: new ListView.separated(
+      child: ListView.separated(
         controller: _scrollController,
         shrinkWrap: true,
         itemCount: transactionsOfAccount.length,
@@ -41,10 +44,38 @@ class _TransactionListState extends State<TransactionList> {
             Divider(height: 1, color: Colors.black12),
         itemBuilder: (BuildContext context, int index) {
           return Card(
-              child: TransactionRow(
-                  transactionsOfAccount[index], widget.appState.allCategories, widget.isEditable));
+              child: TransactionRow(transactionsOfAccount[index],
+                  widget.appState.allCategories, widget.isEditable));
         },
       ),
     ));
   }
+}
+
+List<MoneyTransaction> _getMoneyTransactions(
+    UnmodifiableListView<MoneyTransaction> transactions, int currentAccountId) {
+  /// Here, [currentAccountId] is the outgoingAccount.
+
+  List<MoneyTransaction> transactionsOfAccount = [];
+  for (var transaction in transactions) {
+    bool isAccountPayee = transaction.payeeID < 0;
+
+    bool currentAccountIsPayeeAccount =
+        -transaction.payeeID == currentAccountId;
+    bool currentAccountIsStandardAccount =
+        transaction.accountID == currentAccountId;
+
+    if ((currentAccountIsStandardAccount && !isAccountPayee) ||
+        (currentAccountIsPayeeAccount && isAccountPayee)) {
+      transactionsOfAccount.add(transaction);
+    } else if ((currentAccountIsStandardAccount && isAccountPayee)) {
+      // The transaction is reversed.i.e. removes money from outAccount(accountId)
+      // into inAccount(payeeId)
+      MoneyTransaction negativeAmountTransaction = transaction.copy();
+      negativeAmountTransaction.amount *= -1;
+      transactionsOfAccount.add(negativeAmountTransaction);
+    }
+  }
+
+  return transactionsOfAccount;
 }